refactor(compare): extract Markdown rendering to render.py (Phase A 3/N)
Move generate_markdown_report and _STATUS_ICONS (294 lines, including
nested helpers _render_c7_detail, _render_unmatched_source, _render_group)
into hv_dataqc/compare/render.py.

The renderer is a clean leaf module: it only depends on CheckResult and
md_escape from _common, plus stdlib (datetime, re, collections). It reads
CheckResult.message and detail dicts but doesn't reach into crosswalk or
check internals.

Status icon constant renamed _STATUS_ICONS → STATUS_ICONS (module-public).
Local md_escape alias inside the moved code uses the public name from
_common; no `_md_escape` import retained in the main file (removed in
this commit since nothing else in compare_source_harmonized.py uses it).

Regression-check: COPDGene, ARIC, HCHS — MD + JSON all match. Tests: 79/79
pass.
